move proper function to helper file and disentagle display logic
